Project-Management-system_Ali_Salama/requirements.txt
crewai
streamlit
pydantic
python-dotenv
requests
pysqlite3-binary